+++ src/core/cache/decorators/ttl/index.ts
@@ -5,41 +5,43 @@
  * Released under the MIT license
  * https://github.com/V4Fire/Core/blob/master/LICENSE
  */
+
 /**
  * [[include:core/cache/simple/README.md]]
  * @packageDocumentation
  */
+
 import type Cache from 'core/cache/interface';
 import type { TTLCache, ClearFilter, DecoratorOptions } from 'core/cache/interface';
 
 export * from 'core/cache/interface';
 
-export default function wrapCacheWithTTL<V = unknown, K = string>(cache: Cache<V, K>, ttl?: number): TTLCache<V, K> {
+export default function addTTL<V = unknown, K = string>(cache: Cache<V, K>, ttl?: number): TTLCache<V, K> {
 	const
 		cacheWithTTL: TTLCache<V, K> = Object.create(cache),
-		ttlMemory = new Map<K, number>();
+		ttlTimers = new Map<K, number | NodeJS.Timeout>();
 
-	cacheWithTTL.set = (key: K, value: V, options?: DecoratorOptions) => {
-		if (options?.ttl != null || ttl != null) {
+	cacheWithTTL.set = (key: K, value: V, opts?: DecoratorOptions) => {
+		if (opts?.ttl != null || ttl != null) {
 			const
-				time = options?.ttl ?? ttl;
+				time = opts?.ttl ?? ttl;
 
-			cacheWithTTL.clearTTL(key);
-			ttlMemory.set(key, (<Window['setTimeout']>setTimeout)(() => cacheWithTTL.remove(key), time));
+			cacheWithTTL.removeTTLFrom(key);
+			ttlTimers.set(key, setTimeout(() => cacheWithTTL.remove(key), time));
 		}
 
-		return cache.set(key, value, options);
+		return cache.set(key, value, opts);
 	};
 
 	cacheWithTTL.remove = (key: K) => {
-		cacheWithTTL.clearTTL(key);
+		cacheWithTTL.removeTTLFrom(key);
 		return cache.remove(key);
 	};
 
-	cacheWithTTL.clearTTL = function clearTTL(key: K) {
-		if (ttlMemory.has(key)) {
-			clearTimeout(ttlMemory.get(key));
-			ttlMemory.delete(key);
+	cacheWithTTL.removeTTLFrom = function removeTTLFrom(key: K) {
+		if (ttlTimers.has(key)) {
+			clearTimeout(<number>ttlTimers.get(key));
+			ttlTimers.delete(key);
 		}
 	};
 
@@ -48,7 +50,7 @@ export default function wrapCacheWithTTL<V = unknown, K = string>(cache: Cache<V
 			removed = cache.clear(filter);
 
 		removed.forEach((_, key) => {
-			cacheWithTTL.clearTTL(key);
+			cacheWithTTL.removeTTLFrom(key);
 		});
 
 		return removed;